pyproject.toml
[tool.poetry]
name = "brewbuds"
version = "0.1.0"
description = ""
authors = ["hwstar-1204 <115150798+hwstar-1204@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
pre-commit = "^3.8.0"
django-allauth = "^65.0.2"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
dj-rest-auth = "^6.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pymysql = "^1.1.1"
gunicorn = "^23.0.0"
cryptography = "^43.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
